pyproject.toml
[project]
name = "celeroot"
version = "0.1.0"
description = "System administration through Celery tasks for Debian/Ubuntu systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "click>=8.0.0",
    "croniter>=1.0.0",
    "pyyaml>=6.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pyrefly>=0.35.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
celeroot = "celeroot.__main__:app"
